\name{isNumeric}
\alias{isNumeric}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Is the input a numeric variable
}
\description{
Is the input a numeric variable}
\usage{
isNumeric(constant)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{constant}{
Input
}
}

\value{
TRUE or FALSE}
