\name{lavaan2ram}
\alias{lavaan2ram}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
lavaan output to ram matrices}
\description{
lavaan output to ram matrices}
\usage{
lavaan2ram(fitModel, digits = 2, zero.print = "0", ram.out = TRUE, fit = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fitModel}{
A lavaan model
}
  \item{digits}{
Digits for number print}
  \item{zero.print}{
Format zeros in the matrix
}
  \item{ram.out}{
Whether print ram matrices
}
  \item{fit}{
Whether print fit statistics
}
}
