\name{makeBridgeList}
\alias{makeBridgeList}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate all bridges
}
\description{
Generate all bridges based on Boker, McArdle, & Neale (2002)}
\usage{
makeBridgeList(pathList, spanList)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pathList}{
A path list from the function \code{\link{makePathList}}
}
  \item{spanList}{
A span list from the function \code{\link{makeSpanList}}
}
}
\references{
Boker, S. M., McArdle, J. J. & Neale, M. C. (2002) An algorithm for the hierarchical organization of path diagrams and calculation of components of covariance between variables. Structural Equation Modeling, 9(2), 174-194

Zhang, Z., Hamagami, F., Grimm, K. J., & McArdle, J. J. (2015). Using R package RAMpath for tracing SEM path diagrams and conducting complex longitudinal data analysis. Structural Equation Modeling, 22(1), 132-147. https://doi.org/10.1080/10705511.2014.935257
}
