\name{ram2lavaan}
\alias{ram2lavaan}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
RAM model to lavaan input
}
\description{
RAM matrices to lavaan input}
\usage{
ram2lavaan(model)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{
An ram model
}
}
