\name{ramEffectSE}
\alias{ramEffectSE}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Sobel standard error for a given effect
}
\description{
Sobel standard error for a given effect
}
\usage{
ramEffectSE(object, effect, path=TRUE)}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
An RAM path bridge output}
  \item{effect}{
The effect to calculate se for. It is in the form a > b > c.
}
  \item{path}{
se for the direct and indirect effect.
}
}