\name{ramFit}
\alias{ramFit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Fit a model using lavaan based on ram input
}
\description{
Fit a model using lavaan based on ram input}
\usage{
ramFit(ramModel, data, digits = 3, zero.print = "0", ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ramModel}{
An ram model
}
  \item{data}{
data
}
  \item{digits}{
Digits for print
}
  \item{zero.print}{
Format of zeros
}
  \item{\dots}{
Options for lavaan
}
}
