\name{RAMMatrix}
\alias{RAMMatrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate RAM matrices based on RAM input}
\description{
Generate RAM matrices based on RAM input}
\usage{
RAMMatrix(model)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{
An RAM model}
}
