\name{RAMParseLavaan}
\alias{RAMParseLavaan}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
lavaan to RAM 
}
\description{
lavaan to RAM matrices}
\usage{
RAMParseLavaan(input, manifest, type = 0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{input}{
lavaan input
}
  \item{manifest}{
observed variables
}
  \item{type}{
0: single headed arrow, ...
}
}
