\name{ramPATH-package}
\alias{ramPATH-package}
\alias{ramPATH}
\docType{package}
\title{
To be added
}
\description{
To be added
}
\details{
\tabular{ll}{
Package: \tab ramPATH\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-29\cr
License: \tab What license is it under?\cr
}

}
\author{
Johnny Zhang and Jack McArdle
}
