\name{RAMPathBridge}
\alias{RAMPathBridge}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate path and bridges
}
\description{
Generate path and bridges}
\usage{
RAMPathBridge(RAMmatrix, allbridge = TRUE, indirect = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{RAMmatrix}{
RAM matrices
}
  \item{allbridge}{
Generate all bridges
}
  \item{indirect}{
Generate all indirect effects
}
}
