\name{ramPlot}
\alias{ramPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot the path diagram according to ram path and bridges
}
\description{
Plot the path diagram according to ram path and bridges}
\usage{
ramPlot(pathbridge, file, from, to, type = c("path", "bridge"), size = c(8, 8), node.font = c("Helvetica", 14), edge.font = c("Helvetica", 10), rank.direction = c("LR", "TB"), digits = 2, output.type = c("graphics", "dot"), graphics.fmt = "pdf", dot.options = NULL, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pathbridge}{
Output from the ramPathBridge function
}
  \item{file}{
File name for the generated figures
}
  \item{from}{
from variable
}
  \item{to}{
to variable
}
  \item{type}{
path: to plot the effect; bridge: to plot the bridges
}
  \item{size}{
To be added
}
  \item{node.font}{
To be added
}
  \item{edge.font}{
To be added
}
  \item{rank.direction}{
To be added
}
  \item{digits}{
To be added
}
  \item{output.type}{
To be added
}
  \item{graphics.fmt}{
To be added
}
  \item{dot.options}{
To be added
}
  \item{\dots}{
To be added
}
}
