\name{ramSummary}
\alias{ramSummary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate the total and individual contribution for each path and bridge
}
\description{
Calculate the total and individual contribution for each path and bridge}
\usage{
ramSummary(pathbridge, from, to, type = c("path", "bridge"))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pathbridge}{
Output from the ramPathBridge function
}
  \item{from}{
from variable
}
  \item{to}{
to variable
}
  \item{type}{
path: to plot the effect; bridge: to plot the bridges
}
}
