\name{ramUniquePath}
\alias{ramUniquePath}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get the uniques paths
}
\description{
Get the uniques paths}
\usage{
ramUniquePath(tPathlist)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tPathlist}{
The path list.
}
}
