\name{makeBridgeList}
\alias{makeBridgeList}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate all bridges
}
\description{
Generate all bridges}
\usage{
makeBridgeList(pathList, spanList)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pathList}{
A path list from the function \code{\link{makePathList}}
}
  \item{spanList}{
A span list from the function \code{\link{makeSpanList}}
}
}
