\name{makePathList}
\alias{makePathList}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Make a list of effects
}
\description{
Make a list of effects}
\usage{
makePathList(AMatrix, indirect = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{AMatrix}{
A matrix from the ram matrices
}
  \item{indirect}{
Whether to generate all indirect effects}
}
