\name{makeSpanList}
\alias{makeSpanList}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Make a list of spans
}
\description{
Make a list of spans}
\usage{
makeSpanList(SMatrix)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{SMatrix}{
S matrix from the ram matrices
}
}
