\name{plot.lcs.power}
\alias{plot.lcs.power}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot the power curve for each specified parameter
}
\description{
Plot the power curve for each specified parameter}
\usage{
\method{plot}{lcs.power}(x, parameter, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Output from the \code{powerLCS} function or Output from the \code{powerBLCS} function
}
  \item{parameter}{
parameter to be plotted.
}
  \item{\dots}{
Options for the plot function.
}
}

\references{
Zhang, Z., & Liu, H. (2016). Sample Size Planning for Latent Change Score Models through Monte Carlo Simulation.
}
