\name{ram2lavaan}
\alias{ram2lavaan}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
RAM model to lavaan model
}
\description{
Convert RAM matrix specification to a lavaan model}
\usage{
ram2lavaan(model)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{
An ram model
}
}
\references{
Yves Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. URL http://www.jstatsoft.org/v48/i02/.

Zhang, Z., Hamagami, F., Grimm, K. J., & McArdle, J. J. (2015). Using R package RAMpath for tracing SEM path diagrams and conducting complex longitudinal data analysis. Structural Equation Modeling, 22(1), 132-147. https://doi.org/10.1080/10705511.2014.935257
}