\name{ram2lavaan}
\alias{ram2lavaan}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
RAM model to lavaan model
}
\description{
Convert RAM matrix specification to a lavaan model}
\usage{
ram2lavaan(model)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{
An ram model
}
}
\references{
Yves Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. URL http://www.jstatsoft.org/v48/i02/.

Zhang, Z., Hamagami, F., Grimm, K. J., & McArdle, J. J. (2013). Using R Package RAMpath for Tracing SEM Path Diagrams and Conducting Complex Longitudinal Data Analysis. Structural Equation Modeling.
}