\name{ramLCS}
\alias{ramLCS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Univariate latent change score model
}
\description{
Univariate latent change score model}
\usage{
ramLCS(data, y, timey, ram.out = FALSE, betay, my0, mys, varey, vary0, varys, vary0ys, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
data
}
  \item{y}{
y data
}
  \item{timey}{
time of y
}
  \item{ram.out}{
To be added
}
  \item{betay}{
To be added
}
  \item{my0}{
To be added
}
  \item{mys}{
To be added
}
  \item{varey}{
To be added
}
  \item{vary0}{
To be added
}
  \item{varys}{
To be added
}
  \item{vary0ys}{
To be added
}
  \item{\dots}{
To be added
}
}
