\name{ramMatrix}
\alias{ramMatrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate ram matrices based on ram input}
\description{
Generate ram matrices based on ram input}
\usage{
ramMatrix(model)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{
An ram model}
}
