\name{ramParseLavaan}
\alias{ramParseLavaan}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
lavaan to ram 
}
\description{
lavaan to ram matrices}
\usage{
ramParseLavaan(input, manifest, type = 0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{input}{
lavaan input
}
  \item{manifest}{
observed variables
}
  \item{type}{
0: single headed arrow, ...
}
}
