\name{RAMpath-package}
\alias{RAMpath-package}
\alias{RAMpath}
\docType{package}
\title{
RAMpath for SEM analysis
}
\description{
RAMpath is a rewrite of RAMpath software developed by John McArdle and Steven Boker.
}
\details{
\tabular{ll}{
Package: \tab RAMpath\cr
Type: \tab Package\cr
Version: \tab 0.1.6\cr
Date: \tab 2012-05-29\cr
License: \tab GPL\cr
}

}
\author{
Zhiyong Zhang, Jack McArdle, Aki Hamagami, and Kevin Grimm
Maintainer: Zhiyong Zhang <zhiyongzhang@nd.edu>
}
\references{
Boker, S. M., McArdle, J. J. & Neale, M. C. (2002) An algorithm for the hierarchical organization of path diagrams and calculation of components of covariance between variables. Structural Equation Modeling, 9(2), 174-194

Yves Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. URL http://www.jstatsoft.org/v48/i02/.

Zhang, Z., Hamagami, F., Grimm, K. J., & McArdle, J. J. (2013). Using R Package RAMpath for Tracing SEM Path Diagrams and Conducting Complex Longitudinal Data Analysis. Structural Equation Modeling.
}
