\name{RAMpath-package}
\alias{RAMpath-package}
\alias{RAMpath}
\docType{package}
\title{
RAMpath for SEM analysis
}
\description{
RAMpath is a rewrite of RAMpath software developed by John McArdle and Steven Boker.
}
\details{
\tabular{ll}{
Package: \tab RAMpath\cr
Type: \tab Package\cr
Version: \tab 0.1.5\cr
Date: \tab 2012-05-29\cr
License: \tab GPL\cr
}

}
\author{
Zhiyong Zhang and Jack McArdle
Maintainer: Zhiyong Zhang <zhiyongzhang@nd.edu>
}
