\name{ramPathBridge}
\alias{ramPathBridge}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate path and bridges
}
\description{
Generate path and bridges}
\usage{
ramPathBridge(rammatrix, allbridge = FALSE, indirect = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{rammatrix}{
RAM matrices
}
  \item{allbridge}{
Generate all bridges
}
  \item{indirect}{
Generate all indirect effects
}
}
